@@ -19,7 +19,7 @@ namespace NEventStore.Serialization.Bson {
     // class via a tool like ResGen or Visual Studio.
     // To add or remove a member, edit your .ResX file then rerun ResGen
     // with the /str option, or rebuild your VS project.
-    [global::System.CodeDom.Compiler.GeneratedCodeAttribute("System.Resources.Tools.StronglyTypedResourceBuilder", "16.0.0.0")]
+    [global::System.CodeDom.Compiler.GeneratedCodeAttribute("System.Resources.Tools.StronglyTypedResourceBuilder", "17.0.0.0")]
     [global::System.Diagnostics.DebuggerNonUserCodeAttribute()]
     [global::System.Runtime.CompilerServices.CompilerGeneratedAttribute()]
     internal class Messages {
@@ -66,24 +66,6 @@ namespace NEventStore.Serialization.Bson {
         internal static string DeserializingStream {
             get {
                 return ResourceManager.GetString("DeserializingStream", resourceCulture);
-            }
-        }
-        
-        /// <summary>
-        ///   Looks up a localized string similar to Inspecting text-based stream contents..
-        /// </summary>
-        internal static string InspectingTextStream {
-            get {
-                return ResourceManager.GetString("InspectingTextStream", resourceCulture);
-            }
-        }
-        
-        /// <summary>
-        ///   Looks up a localized string similar to The encryption key must be exactly 16 bytes..
-        /// </summary>
-        internal static string InvalidKeyLength {
-            get {
-                return ResourceManager.GetString("InvalidKeyLength", resourceCulture);
             }
         }
         
