@@ -10,6 +10,7 @@ namespace NEventStore.Serialization.Json
         /// <summary>
         /// Specify we want to use Json serialization using Newtonsoft.Json
         /// </summary>
+        /// <param name="wireup">The persistence wire-up.</param>
         /// <param name="jsonSerializerSettings">
         /// Allows to customize some Serializer options, however some of them will
         /// be under control of this specific implementation and will be overwritten no matter
@@ -20,7 +21,7 @@ namespace NEventStore.Serialization.Json
         /// </param>
         public static SerializationWireup UsingJsonSerialization(
             this PersistenceWireup wireup,
-            JsonSerializerSettings jsonSerializerSettings = null)
+            JsonSerializerSettings? jsonSerializerSettings = null)
         {
             return wireup.UsingCustomSerialization(new JsonSerializer(jsonSerializerSettings));
         }
