@@ -1,5 +1,7 @@
 using MessagePack;
 using MessagePack.Resolvers;
+using Microsoft.Extensions.Logging;
+using NEventStore.Logging;
 
 namespace NEventStore.Serialization.MsgPack
 {
@@ -8,6 +10,7 @@ namespace NEventStore.Serialization.MsgPack
     /// </summary>
     public class MsgPackSerializer : ISerialize
     {
+        private static readonly ILogger Logger = LogFactory.BuildLogger(typeof(MsgPackSerializer));
         /// <summary>
         /// Serializer options
         /// </summary>
@@ -28,7 +31,14 @@ namespace NEventStore.Serialization.MsgPack
         /// <typeparam name="T">Object type</typeparam>
         /// <param name="output">Output stream</param>
         /// <param name="graph">Object to deserialize.</param>
-        public virtual void Serialize<T>(Stream output, T graph) => MessagePackSerializer.Serialize(output, graph, _options);
+        public virtual void Serialize<T>(Stream output, T graph) where T : notnull
+        {
+            if (Logger.IsEnabled(LogLevel.Trace))
+            {
+                Logger.LogTrace(Messages.SerializingGraph, typeof(T));
+            }
+            MessagePackSerializer.Serialize(output, graph, _options);
+        }
 
         /// <summary>
         /// Deserializes an object from stream.
@@ -36,6 +46,13 @@ namespace NEventStore.Serialization.MsgPack
         /// <typeparam name="T">Object type.</typeparam>
         /// <param name="input">Stream input</param>
         /// <returns>Deserialized object</returns>
-        public virtual T Deserialize<T>(Stream input) => MessagePackSerializer.Deserialize<T>(input, _options);
+        public virtual T Deserialize<T>(Stream input)
+        {
+            if (Logger.IsEnabled(LogLevel.Trace))
+            {
+                Logger.LogTrace(Messages.DeserializingStream, typeof(T));
+            }
+            return MessagePackSerializer.Deserialize<T>(input, _options);
+        }
     }
 }
