@@ -16,7 +16,7 @@ namespace NEventStore.Serialization
         /// <typeparam name="T">The type of object to be serialized</typeparam>
         /// <param name="output">The stream into which the serialized object graph should be written.</param>
         /// <param name="graph">The object graph to be serialized.</param>
-        void Serialize<T>(Stream output, T graph);
+        void Serialize<T>(Stream output, T graph) where T : notnull;
 
         /// <summary>
         ///     Deserializes the stream provided and reconstructs the corresponding object graph.
@@ -24,6 +24,6 @@ namespace NEventStore.Serialization
         /// <typeparam name="T">The type of object to be deserialized.</typeparam>
         /// <param name="input">The stream of bytes from which the object will be reconstructed.</param>
         /// <returns>The reconstructed object.</returns>
-        T Deserialize<T>(Stream input);
+        T? Deserialize<T>(Stream input);
     }
 }
